refactor(agents): move codex's plugin installer and claude's path encodings out of the app
Four more allowlist entries, and a bundling trap worth recording.

**codex's plugin installer.** `apps/cli/src/commands/install-codex.ts` was 464
lines with ten exports, all of them about codex's plugin marketplace and its
`~/.codex/config.toml` enable table. None is shared CLI machinery, so it is
`@agentbox/agent-codex/cli` now and the command file is a 47-line shim like the
other per-agent commands.

`resolveDevRepoRoot` is injected rather than imported: locating the user's source
checkout reads the running CLI's own path, which is the app's business, and an
agent package must not reach up into `apps/cli`. Threading a parameter like that
is exactly where things get dropped, so there is a test that fails when
`installCodexPlugin` forwards only `noDev` — the dev marketplace and the skill
symlinks would otherwise stop happening with every build and typecheck still
green. Verified live: `install codex --dry-run` still resolves the local repo.

**claude's path encodings.** `encodeClaudeProjectsDir` / `BOX_WORKSPACE_ENCODED`
(Claude Code's `~/.claude/projects/<encoded>/` scheme) left the shared kit for
`@agentbox/agent-claude`; `BOX_CLAUDE_PLANS_DIR` left `session-teleport/plan.ts`,
which now takes the destination as `boxParentDir` and is agent-neutral.

**The trap.** Both moves invited the obvious follow-up — import the constant back
from `@agentbox/agent-claude/cli` — and both times that broke the HUB build, not
this one: `plan.ts` and `fork.ts` are bundled into the hub, and that entry drags
cli-kit's pty backend in, so esbuild fails on a `.node` binary. Same thing bit
`agent-codex`'s main entry, which `sandbox-docker`'s module registry pulls into
the hub. The installer sits on the `./cli` subpath for that reason, with a
source-level test on the main entry, mutation-checked.

**The guard got narrower, not wider.** Rather than exempting the per-agent
command files wholesale, it now exempts the single
`export const <agent>Command = new Command(` line. Those are correct — the
dispatch tables need literal specifiers — but an agent's LOGIC can no longer hide
beside one, which is what forced the codex installer out in the first place.
Planting a helper in `download-claude.ts` fails it.

Allowlist: 13 -> 6.
